playlists: refactor media actions, fixes #26
Media actions are now created by the PlaylistManager (and will also
be created by eg. the History manager later) instead of the Action
buttons. Data is propagated by wrapping the action event handlers.

Components that render MediaLists now have to pass in a makeActions
closure that returns the React elements to render inside the .Actions
element in a specific row. This is a lot more flexible: actions can
now be different for different contexts, and they can even be
different for different rows. For example, the PlaylistManager shows
the MoveToFirst button on all rows except the first.