playlists: add media to playlists, redux™
Extracted the AddingMenu to a "top-level" component. The
AddToPlaylistAction component opens the AddingMenu; the AddingMenu
fires the addMedia action when necessary. Now have to pass a
position to the AddingMenu. This'll make fixing #29 _much_ easier
later.

The AddToPlaylistAction and Action components were converted to
full React Components, to allow using `ref=` on Actions.

Drag-drop uses the raw addMedia action on drop.

Added the Add To Playlist menu state to the Redux store.