playlists: load more playlist items when scrolling
Is this real life?

This implements a PaginatedList component that requests more pages
when a user scrolls near the edge of the loaded data.

To keep it simple, it requires the component users (only MediaList
for now) to feed it a constant item height which can easily be
compared to the current scroll values. Component users also have to
give it a function to load more pages, of course.

The PlaylistPanel passes the loadPlaylist action function to
MediaList/PaginatedList.

TODO: move the inFlightPlaylists hackiness from
PlaylistActionCreators to the PaginatedList component. The problem
occurs because PaginatedList will happily request the same page
thrice if three scroll events happen closely after another, and
depending on all PaginatedList users to work around that isn't a
very nice thing to do.