booth: fix progress bar when booth is empty
Previously, an empty booth would set the progress bar size
to "Infinity%", which caused browsers to keep it at its previous
position.

This patch moves the progress computation to a selector, which is
where computations are supposed to happen in the first place. It
also deals with media durations of 0 (and thus empty booths) by
defaulting the progress bar width to "0%".