video: fix video size changes resetting the youtube player
When using react-youtube 5, changing the `seek` property also
resets the player. But the `seek` property is continuously
changing as time passes, so the YouTubePlayer component used
a shouldComponentUpdate check to filter for only the prop updates
that _should_ trigger a player reset. However that still meant that
changing _other_ props that should trigger a render of the wrapper
elements, but not of the player itself, would also reset the player.
The videoSize setting is the most obvious example of this.

This commit separates the state changes that _should_ reset the
player into a new YouTubePlayerEmbed component, so that the parent
component can rerender wrapper elements freely, and the Embed
component can only rerender when necessary (i.e. when the song
changes.)