overlays: add <Overlay/> wrapper component
Primarily adds `overflow: hidden` to Overlays, so that the rest of
the app can use `overflow: hidden` in fewer places. This will
allow eg. tooltips in the left part of the app, to flow into the
right part of the app.

Also makes setting the direction (top/bottom) for the animations a
bit simpler.