playlists: fix deselecting playlists (→ selecting search results)
The selectors for selected and active playlists naively merged in
the selected and active media lists, but that doesn't do the right
thing when no playlist is selected/active. The list would be merged
into `null`, resulting in an object with just a `{media}` property
instead of `null`. That then broke `null` checks elsewhere.

This patch no longer attempts to merge playlist items into the
playlist object if there is no playlist object.