history: remove duplicate history entry if it's currently playing
Previously if the history was loaded after an advance came in,
the history panel would contain the current entry twice: once
without stats, from the history response, and once with stats,
from the current play.

This patch removes the current play from the loaded history
entries, so that it'll always only show the entry with stats.