video: preview songs in dialog when double-clicking items
This patch is rather invasive because it needs to work with all
types of media lists.

It adds a new bare Player component to render a player for a video
from any source. This is different from the main Video player,
because the main Video players of all types need to stay in the DOM
at all times to deal with Chrome's autoplay behaviour.

The new bare Player component is used with the new "preview" player
mode to render players into a dialog.