store: add middleware for doing API requests, closes #122
This adds a Redux store middleware that executes API requests. It
is still promise-based, as before, but that's now primarily for
sequencing. A request can be kicked off by dispatching a request
action. The request middleware handles adding the authentication
token to requests, so action creators don't have to do that
manually anymore :)

This also puts the actual request handling code inside a _factory_
so it's configurable (for #217, for example).

A request action looks a bit like this:

```js
{
  type: REQUEST_START,
  payload: {
    method, url, // endpoint
    qs, // query parameters for GET requests
    data // request body for other requests
  },
  meta: {
    onStart, onComplete, onError, // handlers
    id // Unique request ID
  }
}
```

`RequestActionCreators` exports functions to create request actions:

 - `get(url, { qs, on... })`
 - `post(url, data, { on... })`
 - `put(url, data, { on... })`,
 - `del(url, data, { on... })`

The handler functions are all optional but when provided should all
return actions. The middleware dispatches these actions so you can
use them to respond to the request result for example.

Example usage:

```js
function getWaitlist() {
  return get('/waitlist', {
    // If the response shape is the same as the handler parameters,
    // you can just do this:
    onStart: getWaitlistStart,
    // Otherwise you may need to change it a bit first e.e
    onComplete: response => getWaitlistComplete(response.waitlist),
    // or you can inline action creators if you really want to
    // ¯\_(ツ)_/¯
    onError: error => ({
      type: GET_WAITLIST_COMPLETE,
      error: true,
      payload: error
    })
  });
}

store.dispatch(getWaitlist());
```