sources: inject media sources into application in one place
Instead of importing known media sources from the files in various
places, import them once at the top level and inject them into the
application. This will make it possible to build extensions for the
üWave web client to add new media sources.

The store setup now needs to know which media sources are available
so that it knows which reducers to call. (Since media sources can
have their own state.)
The available media sources are passed down the React tree using
context. Currently this is just an object of sourceType→object,
but it's abstracted behind an @injectMediaSources helper so we can
change that once client extensions are a thing.