build: Enable lodash paths.
We use the lodash dotted path `a.b.c` functionality in
`mergeIncludedModels`, so enable it again.

Also changes one use of `find` with a matcher object to a function.

Also uses the Lodash Webpack plugin everywhere, instead of just in
the production build, so it becomes easier to catch these problems
in development.