Flatten invalid nested selectors, closes #639 (#642)
This patch fixes some invalid selectors. CSS nesting doesn't actually
work like this:

```css
.Parent {
  &-child {
    color: red;
  }
}
```

Instead, the `&` is a selector that matches real elements, so here
`color: red` only applies to `.Parent`s that also match `-child`,
which doesn't make a whole lot of sense. It used to work because
the postcss nesting plugin happens to compile it to `.Parent-child`.

Ref: https://tabatkins.github.io/specs/css-nesting/#nest-selector