Fix entering fullscreen, fixes #658 (#659)
* `npm update`

* video: Fix entering fullscreen mode, fixes #658

Entering fullscreen must be done _directly_ resulting from a user
gesture (eg. a click). Previously the Video component would request
fullscreen when it was rerendered with a different `isFullscreen`
prop, which worked because the UI was rerendered synchronously when
the fullscreen button was clicked.

Since #629 state changes do not cause synchronous rerenders anymore,
so the fullscreen request didn't occur as a direct result of the button
click.

This patch instead requests fullscreen immediately when the button is
clicked, and only _then_ sends the ENTER_FULLSCREEN action into the
app's internal state.