build: Add ahead of time gzip, brotli compression
Brotli makes the bundle something like 20% smaller which is really
neat.  Now using `connect-gzip-static` instead of bare `serve-static`,
it wraps it so that `.br` and `.gz` files are served instead of the
uncompressed files.  Only JS, CSS, and some audio and SVG files are
compressed, because gzip-likes don't help much for PNGs. Maybe we can
use zopfliPNG via the image-webpack-loader, I don't know if that's a
thing. `.html` files are also untouched because we need to inject config
into the `index.html` at runtime, so we would always use the
uncompressed version anyway and then gzip it afterwards.