Update to Babel 7
And remove the webpack CLI. Just use gulp instead!
The webpack CLI contains a ton of code to create or change configs but
this repo only uses it to run webpack. That's also done in 10 lines of
code, and it saves a massive dependency tree to do it manually, sooo…

Now we do `$ gulp prod` instead of `$ webpack`.