Add server list.
The about page trigger icon is now always shown. If there is no about
page, the Server list will be the initially active and only item. If
there is an about page, the about page will be active initially, and the
server list can be accessed from the left menu.

The server list is lazy loaded; it has to do markdown parsing to show
previews of room descriptions (rules etc) and the markdown parser is
big.

This uses the same code as https://hub.u-wave.net, via the
[@u-wave/react-server-list](https://github.com/u-wave/react-server-list)
module.