Manually serve the web app manifest.
webpack-pwa-manifest doesn't appear to inject itself into the HTML and
it only handles icon resizing anyway. This patch ditches icon resizing
and makes the middleware handle the manifest. It would be better to have
a static manifest in the future but it needs some config (like the app
name) so I'll wait with that for now.