Initialize votes on load (#1103)
Fixes the vote counters always being 0 on load, even if people already
voted.

Previously this was handled via the ADVANCE action, but the stats were
removed from it a while ago. That was because the stats were only
available on the initial ADVANCE, not subsequent ones, which was
inconsistent.

Now it uses the INIT_STATE action which is more appropriate.