New reconnect logic (#1122)
Replaces reconnecting-websocket with custom logic. Now, when reconnecting, we first hit /now to get the socket token. This also syncs state, which should fix the crashes we get after reconnecting after a long time—usually because a user left in the mean time, but local out-of-date state still had them in the waitlist, or because someone joined in the mean time and sent a chat message, but local state doesn't know who they are.