build: replace postcss-preset-env by specific plugins
only using a handful at this point. postcss-preset-env doesn't seem to
quite apply correctly in this setup, where our build infrastructure uses
postcss 8 while preset-env is still on postcss 7.

want to add it back again later so we don't need to be so mindful about
which features we can use.