Do not hide playlist items when playlist meta is loading (#2973)
When a playlist is in the loading state, that doesn't actually mean that
its items should not be accessible. For example, a playlist is loading
while you're activating it. It doesn't make sense to then show a loading
spinner covering the whole playlist manager. We already show a loading
icon on the playlist name.

With this change, playlist deletion / activation / selection / adding
items don't place a loading spinner on the whole playlist panel. Only
when the playlist is in the loading state *and* doesn't have any items
loaded yet, do we show the whole-screen spinner.